+++ server/src/main/java/org/opensearch/action/admin/cluster/settings/TransportClusterUpdateSettingsAction.java
@@ -64,6 +64,7 @@ import org.opensearch.transport.TransportService;
 import java.io.IOException;
 
 import static org.opensearch.cluster.routing.allocation.decider.ShardsLimitAllocationDecider.CLUSTER_TOTAL_PRIMARY_SHARDS_PER_NODE_SETTING;
+import static org.opensearch.cluster.routing.allocation.decider.ShardsLimitAllocationDecider.CLUSTER_TOTAL_REMOTE_CAPABLE_PRIMARY_SHARDS_PER_NODE_SETTING;
 import static org.opensearch.cluster.service.ClusterManagerTask.CLUSTER_UPDATE_SETTINGS;
 import static org.opensearch.index.remote.RemoteStoreUtils.checkAndFinalizeRemoteStoreMigration;
 
@@ -329,11 +330,16 @@ public class TransportClusterUpdateSettingsAction extends TransportClusterManage
 
     private void validateClusterTotalPrimaryShardsPerNodeSetting(ClusterState currentState, ClusterUpdateSettingsRequest request) {
         if (request.transientSettings().hasValue(CLUSTER_TOTAL_PRIMARY_SHARDS_PER_NODE_SETTING.getKey())
-            || request.persistentSettings().hasValue(CLUSTER_TOTAL_PRIMARY_SHARDS_PER_NODE_SETTING.getKey())) {
+            || request.persistentSettings().hasValue(CLUSTER_TOTAL_PRIMARY_SHARDS_PER_NODE_SETTING.getKey())
+            || request.transientSettings().hasValue(CLUSTER_TOTAL_REMOTE_CAPABLE_PRIMARY_SHARDS_PER_NODE_SETTING.getKey())
+            || request.persistentSettings().hasValue(CLUSTER_TOTAL_REMOTE_CAPABLE_PRIMARY_SHARDS_PER_NODE_SETTING.getKey())) {
 
             Settings settings = Settings.builder().put(request.transientSettings()).put(request.persistentSettings()).build();
 
-            int newValue = CLUSTER_TOTAL_PRIMARY_SHARDS_PER_NODE_SETTING.get(settings);
+            int newValue = Math.max(
+                CLUSTER_TOTAL_PRIMARY_SHARDS_PER_NODE_SETTING.get(settings),
+                CLUSTER_TOTAL_REMOTE_CAPABLE_PRIMARY_SHARDS_PER_NODE_SETTING.get(settings)
+            );
 
             // If default value (-1), no validation needed
             if (newValue == -1) {
@@ -351,6 +357,8 @@ public class TransportClusterUpdateSettingsAction extends TransportClusterManage
                 throw new IllegalArgumentException(
                     "Setting ["
                         + CLUSTER_TOTAL_PRIMARY_SHARDS_PER_NODE_SETTING.getKey()
+                        + "] or ["
+                        + CLUSTER_TOTAL_REMOTE_CAPABLE_PRIMARY_SHARDS_PER_NODE_SETTING.getKey()
                         + "] can only be used with remote store enabled clusters"
                 );
             }